Move param decorators to the package that uses them
