+++ packages/koa-serve/src/index.ts
@@ -9,6 +9,7 @@ export {
   DBOSHTTPMethodInfo,
   DBOSHTTPReg,
   DBOSHTTPRequest,
+  DBOSResponseError,
   RequestIDHeader,
   WorkflowIDHeader,
 } from './dboshttp';
